pyproject.toml
[project]
name = "quiz-app"
version = "0.1.0"
description = "College Quiz Management System — CCE B, MIT Manipal"
requires-python = ">=3.10"

dependencies = [
    "cryptography>=46.0.6",
    "flask>=3.0.0",
    "pymysql>=1.1.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
